+++ cosmetics-web/app/models/email_verification_key.rb
@@ -1,5 +1,5 @@
 class EmailVerificationKey < ApplicationRecord
-  belongs_to :responsible_person
+  belongs_to :contact_person
 
   before_create :set_key
   before_create :set_expires_at
@@ -10,13 +10,6 @@ class EmailVerificationKey < ApplicationRecord
 
   def is_expired?
     expires_at < DateTime.current
-  end
-
-  def self.verify_key_for_responsible_person(responsible_person, key)
-    EmailVerificationKey.find_by!(
-      "responsible_person_id = ? AND key = ?",
-      responsible_person.id, key
-    )
   end
 
 private